Resolved timezone issues related to event editing.
All dates in data feeds or HTML pages are now in the local time of the timezone
specified by the TIMEZONE setting. I've changed this to Europe/London, as
that's the timezone we'll be outputting dates for. It may be useful to change
this to something else for testing.

This is achieved by accessing event start and end dates through
event.start_local() and event.end_local(). We should no longer access the
event.start / event.end directly, as the timezone the datetime is in cannot be
guaranteed. For example, if the start/end has been updated by a form then the
start/end will be in the local timezone. If a specific timezone (e.g. UTC) is
desired, it should be passed to *_local(). See the example in Event's
__unicode__ method.

Note that the iCalandar feed is (and was) outputting dates without any
timezone, i.e. to be interpreted as being in the local time of the calendar
consumer. This should be changed.