Start of server side support for list edit page.
The list page now outputs actual modules, series and events (but only
the titles of events, not any other details). Submitting forms back is
not yet implemented. The view to POST to does not yet exist. When it
does it will work by creating a SeriesEditor for the series being
POSTed.