Optional ?depth=x URL parameter on Thing.get_events()
By default takes value 1 (i.e. retrive only the events for the current Thing).
In administrator calendar view takes value 2 (retrieve events for the current Thing and any Things with this Thing as parent).
May not currently go deeper than 2, though it would not be difficult to add more levels.