Add program to convert Apache access logs into JSON keyed by user
Each user found in the access logs is output as a key in the top level
JSON document, with their value being another dict with a key-val
containing the list of iCal fetches that have been made for their
calendar.

This data will be merged with the output of the other 2 user JSON
producing programs to produce our raw data set.

git-svn-id: https://source.caret.cam.ac.uk/svn/projects/timetables/trunk@21711 59899196-baf2-4bc4-a94d-4a08719e171f