Add year to header & support archival warning message
The "Timetable" header now has the site's academic year under it
(e.g. 2013-2014).
This is implemented by using settings.DEFAULT_ACADEMIC_YEAR.

Also, a new setting: NEXT_YEAR_SITE_URL has been added. If set, a
warning banner will be displayed in the administration section,
warning that the site has been archived and linking to the URL
specified in this setting. Also the word "archive" is added to
the header in both student and admin sections.

git-svn-id: https://source.caret.cam.ac.uk/svn/projects/timetables/trunk@24113 59899196-baf2-4bc4-a94d-4a08719e171f