Fix right click > paste edits not being detected
In the admin editor right click > paste in an event's fields was not marking
the value as changed, so when the series was saved the change would be
reverted.

This was because the paste event was firing before the change to the cell
was actually made, so it didn't think anything had chagned. Deferring the
call to the handler makes it see the change and handle it appropriately.

git-svn-id: https://source.caret.cam.ac.uk/svn/projects/timetables/trunk@24310 59899196-baf2-4bc4-a94d-4a08719e171f