Add manage command for externallinks
Currently it supports exporting external links. This will allow us to
spider all the external links in order to archive them. In the future
we'll implement importing links so that we can update the external
links in bulk to local versions of the files to avoid losing content
when 3rd parties take their content down.

git-svn-id: https://source.caret.cam.ac.uk/svn/projects/timetables/trunk@24440 59899196-baf2-4bc4-a94d-4a08719e171f